Refeactor and reposition
