+++ client/components/App.tsx
@@ -1,24 +1,9 @@
-import React from 'react';
-import Grid from '../components/game-grid';
-import CountdownTimer from '../components/Countdown';
+import { Outlet } from 'react-router-dom';
 
 function App() {
-  const gameDuration = 60000;
-  const [gameStarted, setGameStarted] = React.useState(false);
-
-  const startGame = () => {
-    setGameStarted(true);
-  };
-
   return (
     <>
-      <div className="app">
-        <h1>Aim Trainer</h1>
-        {gameStarted && <CountdownTimer duration={gameDuration} />}
-      </div>
-      <div className="grid-container">
-        <Grid onStartGame={startGame} />
-      </div>
+      <Outlet />
     </>
   );
 }
